test: move scanner and ocr tests to utils
