+++ app/src/test/java/com/github/onlynotesswent/utils/ScannerTest.kt
@@ -1,4 +1,4 @@
-package com.github.onlynotesswent.model.scanner
+package com.github.onlynotesswent.utils
 
 import android.app.Activity
 import android.content.ActivityNotFoundException
@@ -17,7 +17,6 @@ import androidx.activity.result.IntentSenderRequest
 import androidx.activity.result.contract.ActivityResultContract
 import androidx.core.content.FileProvider
 import androidx.core.content.FileProvider.getUriForFile
-import com.github.onlynotesswent.utils.Scanner
 import com.google.android.gms.tasks.OnFailureListener
 import com.google.android.gms.tasks.OnSuccessListener
 import com.google.android.gms.tasks.Task
